requirements.txt
Django>=3.2.5
psycopg2-binary>=2.8.6
django-environ >= 0.4.5
djangorestframework >= 3.12
django-treebeard >= 4.0.1